cleaning up unused handlers/emitters and splitting store codee
