+++ packages/jux-ui/src/model/ExecutionStatus.js
@@ -1,6 +1,10 @@
 
+/**
+ * #API
+ */
 const ExecutionStatus = {
   passed: 'passed',
   failed: 'failed'
 }
+
 export default ExecutionStatus
