+++ packages/jux-ui/src/store/FileStatusType.js
@@ -1,0 +1,10 @@
+
+/**
+ *
+ */
+const FileStatusType = {
+  running: 'running',
+  completed: 'completed'
+}
+
+export default FileStatusType
