+++ packages/jux-ui/src/store/ReporterStatusType.js
@@ -1,0 +1,7 @@
+
+const ReporterStatusType = {
+  running: 'running',
+  idle: 'idle'
+}
+
+export default ReporterStatusType
